Bumpet dittnav-dependencies og flyttet inn kode som har vært fjernet fra dittnav-common-lib.

+++ src/main/kotlin/no/nav/personbruker/dittnav/varselbestiller/config/ApplicationContext.kt
@@ -10,10 +10,10 @@ import no.nav.personbruker.dittnav.common.metrics.MetricsReporter
 import no.nav.personbruker.dittnav.common.metrics.StubMetricsReporter
 import no.nav.personbruker.dittnav.common.metrics.influx.InfluxMetricsReporter
 import no.nav.personbruker.dittnav.common.metrics.influx.SensuConfig
-import no.nav.personbruker.dittnav.common.util.kafka.producer.KafkaProducerWrapper
 import no.nav.personbruker.dittnav.varselbestiller.beskjed.BeskjedEventService
 import no.nav.personbruker.dittnav.varselbestiller.common.database.Database
 import no.nav.personbruker.dittnav.varselbestiller.common.kafka.Consumer
+import no.nav.personbruker.dittnav.varselbestiller.common.kafka.KafkaProducerWrapper
 import no.nav.personbruker.dittnav.varselbestiller.common.kafka.polling.PeriodicConsumerPollingCheck
 import no.nav.personbruker.dittnav.varselbestiller.doknotifikasjon.DoknotifikasjonProducer
 import no.nav.personbruker.dittnav.varselbestiller.doknotifikasjonStopp.DoknotifikasjonStoppProducer
@@ -131,7 +131,7 @@ class ApplicationContext {
             val sensuConfig = SensuConfig(
                     applicationName = environment.applicationName,
                     hostName = environment.sensuHost,
-                    hostPort = environment.sensuPort.toInt(),
+                    hostPort = environment.sensuPort,
                     clusterName = environment.clusterName,
                     namespace = environment.namespace,
                     eventsTopLevelName = "dittnav-varselbestiller",
